微調整

+++ packages/Animal/Staff/Create/UseCase/AnimalCreateUseCaseInput.php
@@ -6,7 +6,7 @@ class AnimalCreateUseCaseInput
 {
     public function __construct(
         private readonly string $name,
-        private readonly float  $width,
+        private readonly int  $width,
     )
     {
     }
@@ -22,7 +22,7 @@ class AnimalCreateUseCaseInput
     /**
      * @return float
      */
-    public function getWidth(): float
+    public function getWidth(): int
     {
         return $this->width;
     }
